third_party/WebKit/LayoutTests/accessibility/table-header-expected.txt

This tests that table headers are exposed for accessibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table.rowCount is 2
PASS tableheader.isValid is true
PASS tableheader.childrenCount is 2
PASS firstHeaderText is 'header 1'
PASS secondHeaderText is 'header 2'
PASS successfullyParsed is true

TEST COMPLETE
